remove exampleTheme, include into .default
